+++ package.json
@@ -13,10 +13,10 @@
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@millicast/sdk": "^0.1.33",
-    "dotenv": "^16.0.3",
     "framer-motion": "^6",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2"
   },
   "devDependencies": {
     "@types/events": "^3.0.0",
@@ -25,6 +25,7 @@
     "@typescript-eslint/eslint-plugin": "^5.40.0",
     "@typescript-eslint/parser": "^5.40.0",
     "@vitejs/plugin-react": "^2.1.0",
+    "dotenv": "^16.0.3",
     "eslint": "^8.25.0",
     "eslint-plugin-react": "^7.31.10",
     "prettier": "^2.7.1",
